refactor(nats): simplify qualification checks
